Add button to copy to clipboard directly as an escape hatch if share sheet fails for some reason
Example: Firefox macOS with `dom.webshare.enabled` set to true in about:config will still throw NotSupportedError as the Web Share API itself is still not supported there yet. (at the time of this commit)

https://caniuse.com/web-share

So in this particular case, fallback to copying to clipboard. However, that will fail too as that wasn't user-operated. So throw up a Copy to Clipboard button that they can click on and that should allow them to copy the game result.