Add changelog dialog
Dialog changes were ported over from 2048-clone so that the changelog can be displayed in a wider width.

`serve.json` also updated for the dev server so that the changelog html can be loaded without it redirecting to the path without .html extension (which would 404). At the same time, dev server still needs to load index.html if '/' path is accessed, so it needs to be added as a path to the `cleanUrls` option.